Moved mapping code to Service due to transaction probs on self-invocat.
According to the Spring Reference Manual:

Note: In proxy mode (which is the default), only 'external' method calls
coming in through the proxy will be intercepted. This means that
'self-invocation', i.e. a method within the target object calling some
other method of the target object, won't lead to an actual transaction
at runtime even if the invoked method is marked with @Transactional!
Consider the use of AspectJ mode (see below) if you expect
self-invocations to be wrapped with transactions as well. In this case,
there won't be a proxy in the first place; instead, the target class
will be 'weaved' (i.e. its byte code will be modified) in order to turn
@Transactional into runtime behavior on any kind of method.

When using proxies, the @Transactional annotation should only be applied
to methods with public visibility. If you do annotate protected, private
or package-visible methods with the @Transactional annotation, no error
will be raised, but the annotated method will not exhibit the configured
transactional settings. Consider the use of AspectJ (see below) if you
need to annotate non-public methods.