Removed unnecessary indexes. Fixed Annotation/Datasource structure.
Name index on Discourse and DiscoursePart weren't compatible with the
change of the name datatype to TEXT. Since the indexes aren't really
necessary, they have been removed.

Adding to "mappedBy" sets is unnecessary and problematic. The
corresponding code in the AnnotationInstance, AnnotationAggregate and
DataSourceAggregate has been changed and the unnecessary add-methods
have been removed.