Handle escaped quotes in coursera files; handle question-response links correctly; sort answers chronologically so they can refer back
(we treat a question as the same type of entity as an answer; they're just contributions
that can be linked.  Coursera treats them as different entities, and that wasn't being
translated properly.  A question is a contribution, and we have to return the contribution id
when it's created and add it to a lookup table so the answer can find it.)